requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.3
django-cors-headers==4.2.0
django-ipware==5.0.0
django-user-agents==0.4.0
djangorestframework==3.14.0
drf-yasg==1.21.6
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.1
Pillow==10.0.0
psycopg2-binary==2.9.6
python-decouple==3.8
pytz==2023.3
PyYAML==6.0
requests==2.31.0
ruamel.yaml==0.17.28
ruamel.yaml.clib==0.2.7
sqlparse==0.4.4
typing_extensions==4.7.1
ua-parser==0.18.0
uritemplate==4.1.1
urllib3==2.0.2
user-agents==2.2.0
whitenoise==6.5.0
